feat: app errors dialog box and own api key support
This commit introduce new dialog with app errors, so instead of printing
it to stderr it will display dialog box with information about error,
for now it supports getting location and fetching weather errors, but we
can add more.

Also we have now ability to set our own api key and use it for future
requests. This should close #11 since it's related to api key.